+++ firestore-bigquery-export/firestore-bigquery-change-tracker/src/logs.ts
@@ -216,12 +216,16 @@ export const bigQueryTableInsertErrors = (
   ]
 ) => {
   logger.warn(`Error when inserting data to table.`);
+
   insertErrors.forEach((error) => {
     logger.warn("ROW DATA JSON:");
     logger.warn(error.row);
-    error.errors.forEach((error) =>
-      logger.warn(`ROW ERROR MESSAGE: ${error.message}`)
-    );
+
+    if (error && error.errors) {
+      error.errors.forEach((error) =>
+        logger.warn(`ROW ERROR MESSAGE: ${error.message}`)
+      );
+    }
   });
 };
 
@@ -258,4 +262,8 @@ export function invalidClustering(fields: string) {
 }
 export const tableCreationError = (table, message) => {
   logger.warn(`Error caught creating table`, message);
+};
+
+export const failedToInitializeWait = (message) => {
+  logger.warn(`Failed while waiting to initialize.`, message);
 };