+++ firestore-bigquery-export/scripts/import/src/index.ts
@@ -23,6 +23,7 @@ import * as inquirer from "inquirer";
 import * as util from "util";
 import * as filenamify from "filenamify";
 import { runMultiThread } from "./run";
+import { resolveWildcardIds } from "./config";
 
 import {
   ChangeType,
@@ -37,7 +38,7 @@ const read = util.promisify(fs.readFile);
 const unlink = util.promisify(fs.unlink);
 
 const BIGQUERY_VALID_CHARACTERS = /^[a-zA-Z0-9_]+$/;
-const FIRESTORE_VALID_CHARACTERS = /^[^\/]+$/;
+const FIRESTORE_VALID_CHARACTERS = /^[^({|})]+$/;
 
 const PROJECT_ID_MAX_CHARS = 6144;
 const FIRESTORE_COLLECTION_NAME_MAX_CHARS = 6144;
@@ -101,8 +102,12 @@ const validateLocation = (value: string) => {
   return index !== -1;
 };
 
+const packageJson = require("../package.json");
+
 program
   .name("fs-bq-import-collection")
+  .description(packageJson.description)
+  .version(packageJson.version)
   .option(
     "--non-interactive",
     "Parse all input from command line flags instead of prompting the caller.",
@@ -279,6 +284,7 @@ const run = async (): Promise<number> => {
     tableId: tableId,
     datasetId: datasetId,
     datasetLocation,
+    wildcardIds: queryCollectionGroup,
   });
 
   console.log(
@@ -298,10 +304,7 @@ const run = async (): Promise<number> => {
     `/from-${formattedPath}-to-${projectId}_${datasetId}_${rawChangeLogName}`;
   if (await exists(cursorPositionFile)) {
     let cursorDocumentId = (await read(cursorPositionFile)).toString();
-    cursor = await firebase
-      .firestore()
-      .doc(cursorDocumentId)
-      .get();
+    cursor = await firebase.firestore().doc(cursorDocumentId).get();
     console.log(
       `Resuming import of Cloud Firestore Collection ${sourceCollectionPath} ${
         queryCollectionGroup ? " (via a Collection Group query)" : ""
@@ -339,10 +342,9 @@ const run = async (): Promise<number> => {
       return {
         timestamp: new Date().toISOString(), // epoch
         operation: ChangeType.IMPORT,
-        documentName: `projects/${projectId}/databases/${FIRESTORE_DEFAULT_DATABASE}/documents/${
-          snapshot.ref.path
-        }`,
+        documentName: `projects/${projectId}/databases/${FIRESTORE_DEFAULT_DATABASE}/documents/${snapshot.ref.path}`,
         documentId: snapshot.id,
+        pathParams: resolveWildcardIds(sourceCollectionPath, snapshot.ref.path),
         eventId: "",
         data: snapshot.data(),
       };